+++ release-notes.md
@@ -3,6 +3,7 @@
 * Add GM-only field to prevent items from being added to card decks ([#44](https://github.com/wrycu/paranoia/issues/44))
 * Harden card state tracking ([#46](https://github.com/wrycu/paranoia/issues/46))
 * Overall moved console logs to a debugging setting and moved some messaged to the UI ([#50](https://github.com/wrycu/paranoia/issues/50))
+* Fix for re-roller using initial attr / skill combo instead of selected ([#53](https://github.com/wrycu/paranoia/issues/53))
 
 `1.6.0` - 2023-09-12
 * Update token names to display full name (`NAME-CLEARANCE-HOME SECTOR-CLONE #`) ([#35](https://github.com/wrycu/paranoia/issues/35))